% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{formatData}
\alias{formatData}
\title{formatData}
\usage{
formatData(inData, inclPhenology = TRUE, minSite = 1, maxSite = 999)
}
\arguments{
\item{inData}{A dataset produced by the simulations}

\item{inclPhenology}{should the model account for seasonal variation?}

\item{minSite}{the threshold minimum number of sites for a species to be considered for modelling}

\item{maxSite}{defines a limit on the number of sites in the database}
}
\value{
list of two data frames
}
\description{
formatData
}
\details{
Formats data ready for Nimble
}
