% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{formatData}
\alias{formatData}
\title{formatData}
\usage{
formatData(
  inData,
  format = "Nimble",
  ListLen = NULL,
  inclPhenology = TRUE,
  minYrPerSite = 2,
  minSite = 1,
  maxSite = 999
)
}
\arguments{
\item{inData}{A dataset with columns "species", "siteID", "survey" and "year"}

\item{format}{Either "Nimble" (default) or "spOcc"}

\item{ListLen}{should be included as continuous ("Cont"), categorical ("Cat") or excluded (NULL)}

\item{inclPhenology}{should the model account for seasonal variation?}

\item{minYrPerSite}{the minimum number of years with data for a site to be included (defaults to 2, as in sparta)}

\item{minSite}{the threshold minimum number of sites for a species to be considered for modelling}

\item{maxSite}{defines a limit on the number of sites in the database}
}
\value{
list of two data frames
}
\description{
formatData
}
\details{
Formats data ready for Nimble
}
