% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseData.R
\name{summariseData}
\alias{summariseData}
\title{summariseData}
\usage{
summariseData(obsData, dataConstants)
}
\arguments{
\item{obsData}{dataframe produced by ProcessSimDatFile()}

\item{dataConstants}{dataframe produced by ProcessSimDatFile()}
}
\value{
summary statistics
}
\description{
summariseData
}
\details{
Summary statistics on the input data
}
